Move cron execution helpers out of route exports
